+++ src/pages/Tools/FindTable/FindTable.tsx
@@ -6,7 +6,7 @@ import { calculateCP, predictCPList, predictStat } from '../../../util/Calculate
 import DataTable from 'react-data-table-component';
 import data from '../../../data/cp_multiplier.json';
 
-import '../../../components/Select/Find/Tools.css';
+import '../../../components/Select/Find/Form.css';
 import { useSnackbar } from 'notistack';
 import { Box, Rating } from '@mui/material';
 import Find from '../../../components/Select/Find/Find';
