+++ src/store/models/searching.model.ts
@@ -1,9 +1,6 @@
 export interface SearchingModel {
-  searching?: string;
-  data: {
-    id: number;
-    name?: string;
-    form?: string;
-    timestamp: Date;
-  };
+  id: number;
+  name?: string;
+  form?: string;
+  timestamp: Date;
 }
